Move TableDataManager to segment-local to unblock partial upsert feature
