+++ pinot-core/src/main/java/org/apache/pinot/core/data/manager/offline/DimensionTableDataManager.java
@@ -30,7 +30,7 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
 import javax.annotation.concurrent.GuardedBy;
 import javax.annotation.concurrent.ThreadSafe;
 import org.apache.pinot.common.metadata.ZKMetadataProvider;
-import org.apache.pinot.core.data.manager.SegmentDataManager;
+import org.apache.pinot.segment.local.data.manager.SegmentDataManager;
 import org.apache.pinot.segment.local.segment.index.loader.IndexLoadingConfig;
 import org.apache.pinot.segment.local.segment.readers.PinotSegmentRecordReader;
 import org.apache.pinot.segment.spi.IndexSegment;
@@ -66,7 +66,8 @@ public class DimensionTableDataManager extends OfflineTableDataManager {
   }
 
   @VisibleForTesting
-  public static DimensionTableDataManager registerDimensionTable(String tableNameWithType, DimensionTableDataManager instance) {
+  public static DimensionTableDataManager registerDimensionTable(String tableNameWithType,
+      DimensionTableDataManager instance) {
     return _instances.computeIfAbsent(tableNameWithType, k -> instance);
   }
 
@@ -116,8 +117,8 @@ public class DimensionTableDataManager extends OfflineTableDataManager {
       _logger.info("Successfully removed segment {} and reloaded lookup table: {}", segmentName, getTableName());
     } catch (Exception e) {
       throw new RuntimeException(String
-          .format("Error reloading lookup table after segment remove '%s' for table: '%s'", segmentName, getTableName()),
-          e);
+          .format("Error reloading lookup table after segment remove '%s' for table: '%s'", segmentName,
+              getTableName()), e);
     }
   }
