--- pinot-core/src/test/resources/data/test_null_value_vector_data.json
@@ -1,12 +1,0 @@
-{
-  "description" : "6\"mlc\u0015x[9jA\u001DT",
-  "secondsSinceEpoch": 1567205394
-}
-{
-  "cityid" : -2053105565,
-  "description" : "q\u0012=cLp{-PDa\fT",
-  "signup_email" : {
-    "string" : "fI\u00194u@WI\f\u001Dt"
-  },
-  "secondsSinceEpoch": 1567205394
-}
