--- pinot-core/src/test/resources/data/test_upsert_schema.json
@@ -1,21 +1,0 @@
-{
-  "schemaName": "test_pinot_table",
-  "dimensionFieldSpecs": [
-    {
-      "name": "event_id",
-      "dataType": "STRING"
-    },
-    {
-      "name": "description",
-      "dataType": "STRING"
-    }
-  ],
-  "timeFieldSpec": {
-    "incomingGranularitySpec": {
-      "name": "secondsSinceEpoch",
-      "dataType": "LONG",
-      "timeType": "SECONDS"
-    }
-  },
-  "primaryKeyColumns": ["event_id"]
-}
